Preserve JavaScript placeholder precedence and control regex contexts
