--- .github/workflows/run-fix302-round7.yml
@@ -1,25 +1,0 @@
-name: Run Codex round seven fixes
-on:
-  push:
-    branches:
-      - security/javascript-placeholder-bindings
-permissions:
-  contents: write
-jobs:
-  patch:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: security/javascript-placeholder-bindings
-      - name: Apply fixes
-        run: python3 .github/scripts/fix302-round7.py
-      - name: Commit fixes
-        run: |
-          rm -f .github/scripts/fix302-round7.py .github/workflows/run-fix302-round7.yml
-          git config user.name 'Ben'
-          git config user.email 'benbergen12@gmail.com'
-          git add -A
-          git commit -m 'Preserve JavaScript placeholder precedence and control regex contexts'
-          git push origin HEAD:security/javascript-placeholder-bindings
