+++ AdvancedCore/src/main/java/com/bencodez/advancedcore/api/javascript/JavascriptPlaceholderBinder.java
@@ -132,15 +132,10 @@ public final class JavascriptPlaceholderBinder {
 
     private static String resolve(String token, OfflinePlayer player, Map<String, String> placeholders) {
         AdvancedCorePlugin plugin = AdvancedCorePlugin.getInstance();
-        // PlaceholderAPI uses percent-delimited placeholders. Brace-delimited tokens
-        // are AdvancedCore's legacy custom placeholder form and are resolved below.
-        if (token.startsWith("%") && player != null && plugin != null && plugin.isPlaceHolderAPIEnabled()) {
-            String resolved = PlaceholderAPI.setPlaceholders(player, token);
-            if (resolved != null && !resolved.equals(token)) {
-                return resolved;
-            }
-        }
 
+        // Preserve the historical replacement order: AdvancedCore custom/reward
+        // placeholders win name collisions, then PlaceholderAPI is applied to the
+        // selected custom value so custom placeholders may themselves contain PAPI.
         if (placeholders != null) {
             String name = token.substring(1, token.length() - 1);
             for (Entry<String, String> entry : placeholders.entrySet()) {
@@ -154,6 +149,15 @@ public final class JavascriptPlaceholderBinder {
                     }
                     return value;
                 }
+            }
+        }
+
+        // Only consult PlaceholderAPI for the original token when no custom
+        // placeholder with the same name was supplied.
+        if (token.startsWith("%") && player != null && plugin != null && plugin.isPlaceHolderAPIEnabled()) {
+            String resolved = PlaceholderAPI.setPlaceholders(player, token);
+            if (resolved != null && !resolved.equals(token)) {
+                return resolved;
             }
         }
         return token;
@@ -466,6 +470,9 @@ public final class JavascriptPlaceholderBinder {
             if ("([{:;,=!?&|+-*%^~<>".indexOf(previous) >= 0) {
                 return true;
             }
+            if (previous == ')' && closesControlStatementHead(source, previousIndex)) {
+                return true;
+            }
 
             if (Character.isJavaIdentifierPart(previous)) {
                 int end = previousIndex + 1;
@@ -481,6 +488,74 @@ public final class JavascriptPlaceholderBinder {
                         || word.equals("new");
             }
             return false;
+        }
+
+        private static boolean closesControlStatementHead(String source, int closeParen) {
+            List<Integer> openingParens = new ArrayList<>();
+            for (int i = 0; i <= closeParen; i++) {
+                char current = source.charAt(i);
+                if (current == '\'' || current == '"') {
+                    i = skipQuotedLiteral(source, i, closeParen + 1, current);
+                    continue;
+                }
+                if (current == '`') {
+                    i = skipTemplateLiteral(source, i, closeParen + 1);
+                    continue;
+                }
+                if (current == '/' && canStartRegex(source, i)) {
+                    int regexEnd = skipRegexLiteral(source, i, closeParen + 1);
+                    if (regexEnd > i) {
+                        i = regexEnd;
+                        continue;
+                    }
+                }
+                if (current == '(') {
+                    openingParens.add(i);
+                } else if (current == ')') {
+                    if (openingParens.isEmpty()) {
+                        return false;
+                    }
+                    int openingParen = openingParens.remove(openingParens.size() - 1);
+                    if (i == closeParen) {
+                        return isControlKeywordBefore(source, openingParen);
+                    }
+                }
+            }
+            return false;
+        }
+
+        private static boolean isControlKeywordBefore(String source, int openingParen) {
+            int end = openingParen - 1;
+            while (end >= 0 && Character.isWhitespace(source.charAt(end))) {
+                end--;
+            }
+            if (end < 0 || !Character.isJavaIdentifierPart(source.charAt(end))) {
+                return false;
+            }
+
+            int start = end;
+            while (start >= 0 && Character.isJavaIdentifierPart(source.charAt(start))) {
+                start--;
+            }
+            String word = source.substring(start + 1, end + 1);
+            if (word.equals("if") || word.equals("while") || word.equals("for") || word.equals("with")
+                    || word.equals("switch") || word.equals("catch")) {
+                return true;
+            }
+
+            // Modern JavaScript may use `for await (...)`.
+            if (!word.equals("await")) {
+                return false;
+            }
+            end = start;
+            while (end >= 0 && Character.isWhitespace(source.charAt(end))) {
+                end--;
+            }
+            start = end;
+            while (start >= 0 && Character.isJavaIdentifierPart(source.charAt(start))) {
+                start--;
+            }
+            return end >= 0 && source.substring(start + 1, end + 1).equals("for");
         }
 
         private static void addPatternRanges(String source, Pattern pattern, List<Range> target,
